Detect if task is part of graph vs group
mozilla-taskcluster attempts to detect when a task is part of a graph
scheduled through the task graph scheduler.  If so, then the reruns
property is evaluated for the task if if the task has failed and will be
auto-rerun, then mozilla-taskcluster will not mark the task as failed
and the next run will cause the previous run to be marked as retried.

mozilla-taskcluster uses the scheduler ID to determine if this was
scheduled through the scheduler, however, during the transition to using
task dependencies via the taskcluster-queue, this scheduler ID has
remained the same.  If looking up the task fails, assume that this task
is part of the new task dependency system within taskcluster-queue.