Only attempt to retrigger one time
This is to limit the out of control effect when retriggering for some reasons fails at some point in the process.  Before when things were scheduled through the scheduler it was better, but now that each task has it's own createTask event on the queue, somewhere in there things could fail, causing some tasks to be created and another retrigger attempted.